src/vpaljwt/requirements/base.txt
click
pyjwt[crypto]

